refactor: Extract truncate tests to separate test file
- Create new test file: crates/vibesql-executor/tests/truncate_tests.rs (591 lines)
- Add test declaration to Cargo.toml
- Remove #[cfg(test)] module from truncate/mod.rs (reduced from 723 to 139 lines)
- Update imports to use vibesql_executor:: instead of crate::

Benefits:
- 81% reduction in module file size (584 lines moved)
- Follows existing pattern of 39 other executor test files
- Improves code navigation and compilation efficiency
- Maintains all 14 truncate tests with no functional changes
